common helpers test
