+++ aws/logs_monitoring_go/internal/parsing/helpers_test.go
@@ -1,0 +1,43 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2026-Present Datadog, Inc.
+
+package parsing
+
+import (
+	"bytes"
+	"compress/gzip"
+	"context"
+	"encoding/json"
+	"testing"
+
+	"github.com/aws/aws-lambda-go/lambdacontext"
+)
+
+const testARN = "arn:aws:lambda:us-east-1:123456789012:function:forwarder"
+
+var testLambdaCtx = lambdacontext.NewContext(context.Background(), &lambdacontext.LambdaContext{
+	InvokedFunctionArn: testARN,
+})
+
+func mustGzipJSON(t *testing.T, v any) []byte {
+	t.Helper()
+
+	raw, err := json.Marshal(v)
+	if err != nil {
+		t.Fatalf("marshal: %v", err)
+	}
+
+	var buf bytes.Buffer
+	w := gzip.NewWriter(&buf)
+	if _, err := w.Write(raw); err != nil {
+		t.Fatalf("gzip write: %v", err)
+	}
+
+	if err := w.Close(); err != nil {
+		t.Fatalf("gzip close: %v", err)
+	}
+
+	return buf.Bytes()
+}
